optionally receving destination operand in expression evaluation functions
sometimes it is better if we can specify the destination operand for an instruction in order to optimize temporary generation. one example is using ABI-specific return register when returning from a function. Instead of generating a temporary for the function's return value(or expression evaluation result), we can directly specify the IROperand::Return to optimize out the intermediate container